refactor: split Lovelace card builders
Move Lovelace card primitive and code-slot builders into focused sibling
modules while keeping registry-backed entity resolution in lovelace.py. The
public generation imports used by integration callers stay anchored there.

Closes #771.
