+++ custom_components/keymaster/lovelace.py
@@ -26,9 +26,10 @@ from .const import (
     CONF_PARENT_ENTRY_ID,
     CONF_SLOTS,
     CONF_START,
-    DAY_NAMES,
     DOMAIN,
 )
+from .lovelace_cards import _generate_lock_badges
+from .lovelace_code_slots import _generate_child_code_slot_dict, _generate_code_slot_dict
 
 _LOGGER: logging.Logger = logging.getLogger(__name__)
 
@@ -381,510 +382,3 @@ def _get_entity_id(
         )
     # If not found in registry, assume it's already a complete entity ID
     return entity_id or prop
-
-
-DIVIDER_CARD = {"type": "divider"}
-
-
-def _generate_entity_card_ll_config(
-    code_slot_num: int,
-    domain: str,
-    key: str,
-    name: str,
-    *,
-    parent: bool = False,
-    type_: str | None = None,
-    tap_action: str = "none",
-    icon: str | None = None,
-) -> MutableMapping[str, Any]:
-    """Generate entity configuration for use in Lovelace cards."""
-    prefix = "parent." if parent else ""
-    entity = f"{prefix}{domain}.code_slots:{code_slot_num}.{key}"
-    data: MutableMapping[str, Any] = {
-        "entity": entity,
-        "name": name,
-        "tap_action": {"action": tap_action},
-        "hold_action": {"action": "none"},
-        "double_tap_action": {"action": "none"},
-    }
-    if type_:
-        data["type"] = type_
-    if icon is not None:
-        data["icon"] = icon
-    return data
-
-
-def _generate_badge_ll_config(
-    entity: str | None,
-    name: str,
-    *,
-    visibility: bool = False,
-    tap_action: str | None = "none",
-    show_name: bool = False,
-    visibility_conditions: list[MutableMapping[str, Any]] | None = None,
-) -> MutableMapping[str, Any]:
-    """Generate Lovelace config for a badge."""
-    data: MutableMapping[str, Any] = {
-        "type": "entity",
-        "show_name": show_name,
-        "color": "",
-    }
-    if tap_action is not None:
-        data["tap_action"] = {"action": tap_action}
-    if show_name:
-        data["name"] = name
-    if entity:
-        data["entity"] = entity
-    if visibility_conditions:
-        data["visibility"] = visibility_conditions
-    elif visibility:
-        data["visibility"] = [
-            {
-                "condition": "state",
-                "entity": "switch.autolock_enabled",
-                "state": "on",
-            }
-        ]
-    return data
-
-
-def _generate_conditional_card_ll_config(
-    code_slot_num: int,
-    domain: str,
-    key: str,
-    name: str,
-    conditions: list[MutableMapping[str, Any]],
-    *,
-    parent: bool = False,
-    type_: str | None = None,
-    tap_action: str = "none",
-    icon: str | None = None,
-) -> MutableMapping[str, Any]:
-    """Generate Lovelace config for a `conditional` card."""
-    return {
-        "type": "conditional",
-        "conditions": conditions,
-        "row": _generate_entity_card_ll_config(
-            code_slot_num,
-            domain,
-            key,
-            name,
-            parent=parent,
-            type_=type_,
-            tap_action=tap_action,
-            icon=icon,
-        ),
-    }
-
-
-def _generate_state_condition(
-    code_slot_num: int,
-    key: str,
-    state: str = "on",
-    parent: bool = False,
-    needs_type: bool = False,
-) -> MutableMapping[str, Any]:
-    """Return the condition for an entity state."""
-    prefix = "parent." if parent else ""
-    data = {
-        "entity": f"{prefix}switch.code_slots:{code_slot_num}.{key}",
-        "state": state,
-    }
-    if needs_type:
-        data["condition"] = "state"
-    return data
-
-
-def _generate_code_slot_conditional_entities_card_ll_config(
-    code_slot_num: int,
-    advanced_date_range: bool,
-    advanced_day_of_week: bool,
-    child: bool = False,
-) -> MutableMapping[str, Any]:
-    """Build the conditional entities card for the code slot."""
-    entities: list[MutableMapping[str, Any]] = [
-        _generate_entity_card_ll_config(code_slot_num, "text", "name", "Name"),
-        _generate_entity_card_ll_config(code_slot_num, "text", "pin", "PIN"),
-        DIVIDER_CARD,
-        _generate_entity_card_ll_config(code_slot_num, "switch", "enabled", "Enabled"),
-        _generate_entity_card_ll_config(code_slot_num, "binary_sensor", "active", "Active"),
-        _generate_entity_card_ll_config(code_slot_num, "event", "last_used", "Last Used"),
-        _generate_entity_card_ll_config(code_slot_num, "sensor", "synced", "Sync Status"),
-        *(
-            (
-                _generate_entity_card_ll_config(
-                    code_slot_num, "switch", "override_parent", "Override Parent"
-                ),
-            )
-            if child
-            else ()
-        ),
-        _generate_entity_card_ll_config(code_slot_num, "switch", "notifications", "Notifications"),
-        DIVIDER_CARD,
-        _generate_entity_card_ll_config(
-            code_slot_num, "switch", "accesslimit_count_enabled", "Limit by Number of Uses"
-        ),
-        _generate_conditional_card_ll_config(
-            code_slot_num,
-            "number",
-            "accesslimit_count",
-            "Uses Remaining",
-            [_generate_state_condition(code_slot_num, "accesslimit_count_enabled")],
-        ),
-        *(_generate_date_range_entities(code_slot_num) if advanced_date_range else ()),
-        *(_generate_dow_entities(code_slot_num) if advanced_day_of_week else ()),
-        DIVIDER_CARD,
-        _generate_entity_card_ll_config(code_slot_num, "button", "reset", "Reset Slot"),
-    ]
-
-    return {
-        "type": "conditional",
-        "conditions": [],
-        "card": {
-            "type": "entities",
-            "show_header_toggle": False,
-            "state_color": True,
-            "entities": entities,
-        },
-    }
-
-
-def _generate_header_ll_config(code_slot_num: int) -> MutableMapping[str, Any]:
-    """Generate Lovelace config for a heading card."""
-    return {"type": "heading", "heading": f"Code Slot {code_slot_num}", "heading_style": "title"}
-
-
-def _generate_code_slot_dict(
-    code_slot_num: int,
-    advanced_date_range: bool,
-    advanced_day_of_week: bool,
-    child: bool = False,
-) -> MutableMapping[str, Any]:
-    """Build the dict for the code slot."""
-    return {
-        "type": "grid",
-        "cards": [
-            _generate_header_ll_config(code_slot_num),
-            _generate_code_slot_conditional_entities_card_ll_config(
-                code_slot_num, advanced_date_range, advanced_day_of_week, child=child
-            ),
-        ],
-    }
-
-
-def _generate_lock_badges(
-    lock_entity: str,
-    door_sensor: str | None = None,
-    battery_entity: str | None = None,
-    child: bool = False,
-) -> list[MutableMapping[str, Any]]:
-    """Generate the Lovelace badges configuration for a keymaster lock."""
-    door = door_sensor is not None
-    battery = battery_entity is not None
-    badges = [
-        _generate_badge_ll_config(
-            entity, name, visibility=visibility, show_name=show_name, tap_action=tap_action
-        )
-        for entity, name, visibility, show_name, tap_action, condition in (
-            ("sensor.lock_name", "Lock Name", False, False, "none", True),
-            ("sensor.parent_name", "Parent Lock", False, True, "none", child),
-            ("binary_sensor.connected", "Connected", False, False, "none", True),
-            ("switch.lock_notifications", "Lock Notifications", False, True, "toggle", True),
-            ("switch.door_notifications", "Door Notifications", False, True, "toggle", door),
-            (lock_entity, "Lock", False, True, "toggle", True),
-            (door_sensor, "Door", False, True, "none", door),
-            (battery_entity, "Battery", False, True, "none", battery),
-            ("switch.autolock_enabled", "Auto Lock", False, True, "toggle", True),
-            ("switch.retry_lock", "Retry Lock", True, True, "toggle", door),
-            ("number.autolock_min_day", "Day Auto Lock", True, True, None, True),
-            ("number.autolock_min_night", "Night Auto Lock", True, True, None, True),
-        )
-        if condition
-    ]
-    badges.append(
-        _generate_badge_ll_config(
-            entity="sensor.autolock_timer",
-            name="Auto Lock Timer",
-            show_name=True,
-            tap_action="none",
-            visibility_conditions=[
-                {
-                    "condition": "state",
-                    "entity": "sensor.autolock_timer",
-                    "state_not": "unknown",
-                },
-                {
-                    "condition": "state",
-                    "entity": "sensor.autolock_timer",
-                    "state_not": "unavailable",
-                },
-            ],
-        )
-    )
-    return badges
-
-
-def _generate_dow_entities(
-    code_slot_num: int, parent: bool = False
-) -> list[MutableMapping[str, Any]]:
-    """Build the day of week entities for the code slot."""
-    _dow_prefix = "accesslimit_day_of_week"
-    type_ = "simple-entity" if parent else None
-    # Name differs for parent vs non-parent views
-    limit_by_time_name = "Limit by Time" if parent else "Limit by Time of Day"
-    return [
-        *([] if parent else [DIVIDER_CARD]),
-        _generate_entity_card_ll_config(
-            code_slot_num,
-            "switch",
-            f"{_dow_prefix}_enabled",
-            "Limit by Day of Week",
-            parent=parent,
-            type_=type_,
-        ),
-        # Generate conditional cards for each day of week.
-        # num_conditions controls visibility nesting via [:num_conditions] slice:
-        #   1 = show when DOW enabled
-        #   2 = show when DOW enabled AND this day enabled
-        #   3 = show when DOW enabled AND this day enabled AND limit_by_time on
-        *(
-            _generate_conditional_card_ll_config(
-                code_slot_num,
-                domain,
-                f"{_dow_prefix}:{dow_num}.{key}",
-                name,
-                [
-                    _generate_state_condition(
-                        code_slot_num, f"{_dow_prefix}{suffix}", parent=parent
-                    )
-                    for suffix in (
-                        "_enabled",
-                        f":{dow_num}.dow_enabled",
-                        f":{dow_num}.limit_by_time",
-                    )[:num_conditions]
-                ],
-                parent=parent,
-                type_=type_,
-            )
-            for dow_num, dow in enumerate(DAY_NAMES)
-            for domain, key, name, num_conditions in (
-                ("switch", "dow_enabled", dow, 1),
-                ("switch", "limit_by_time", limit_by_time_name, 2),
-                ("switch", "include_exclude", "Include (On)/Exclude (Off) Time", 3),
-                ("time", "time_start", "Start Time", 3),
-                ("time", "time_end", "End Time", 3),
-            )
-        ),
-    ]
-
-
-def _generate_date_range_entities(
-    code_slot_num: int, parent: bool = False
-) -> list[MutableMapping[str, Any]]:
-    """Build the date range entities for the code slot."""
-    type_ = "simple-entity" if parent else None
-    # For non-parent datetime rows, use simple-entity with more-info tap
-    # to avoid the inline datetime picker overflowing card boundaries.
-    datetime_type = "simple-entity"
-    datetime_tap = "none" if parent else "more-info"
-    datetime_icon = None if parent else "mdi:pencil"
-    return [
-        *([] if parent else [DIVIDER_CARD]),
-        _generate_entity_card_ll_config(
-            code_slot_num,
-            "switch",
-            "accesslimit_date_range_enabled",
-            "Limit by Date Range",
-            parent=parent,
-            type_=type_,
-        ),
-        _generate_conditional_card_ll_config(
-            code_slot_num,
-            "datetime",
-            "accesslimit_date_range_start",
-            "Date Range Start",
-            [
-                _generate_state_condition(
-                    code_slot_num, "accesslimit_date_range_enabled", parent=parent
-                )
-            ],
-            parent=parent,
-            type_=datetime_type,
-            tap_action=datetime_tap,
-            icon=datetime_icon,
-        ),
-        _generate_conditional_card_ll_config(
-            code_slot_num,
-            "datetime",
-            "accesslimit_date_range_end",
-            "Date Range End",
-            [
-                _generate_state_condition(
-                    code_slot_num, "accesslimit_date_range_enabled", parent=parent
-                )
-            ],
-            parent=parent,
-            type_=datetime_type,
-            tap_action=datetime_tap,
-            icon=datetime_icon,
-        ),
-    ]
-
-
-def _generate_parent_view_entities(
-    code_slot_num: int,
-    advanced_date_range: bool,
-    advanced_day_of_week: bool,
-    parent_pin_entity_id: str | None = None,
-) -> list[MutableMapping[str, Any]]:
-    """Build entity rows for a child lock parent-view card."""
-    entities: list[MutableMapping[str, Any]] = [
-        _generate_entity_card_ll_config(
-            code_slot_num, "text", "name", "Name", parent=True, type_="simple-entity"
-        ),
-    ]
-
-    if not parent_pin_entity_id:
-        entities.append(
-            _generate_entity_card_ll_config(
-                code_slot_num, "text", "pin", "PIN", parent=True, type_="simple-entity"
-            )
-        )
-
-    entities.extend(
-        [
-            _generate_entity_card_ll_config(
-                code_slot_num, "switch", "enabled", "Enabled", parent=True, type_="simple-entity"
-            ),
-            _generate_entity_card_ll_config(code_slot_num, "binary_sensor", "active", "Active"),
-            _generate_entity_card_ll_config(code_slot_num, "event", "last_used", "Last Used"),
-            _generate_entity_card_ll_config(code_slot_num, "sensor", "synced", "Sync Status"),
-            _generate_entity_card_ll_config(
-                code_slot_num, "switch", "override_parent", "Override Parent"
-            ),
-            _generate_entity_card_ll_config(
-                code_slot_num, "switch", "notifications", "Notifications"
-            ),
-            _generate_entity_card_ll_config(
-                code_slot_num,
-                "switch",
-                "accesslimit_count_enabled",
-                "Limit by Number of Uses",
-                parent=True,
-                type_="simple-entity",
-            ),
-            _generate_conditional_card_ll_config(
-                code_slot_num,
-                "number",
-                "accesslimit_count",
-                "Uses Remaining",
-                [
-                    _generate_state_condition(
-                        code_slot_num, "accesslimit_count_enabled", parent=True
-                    )
-                ],
-                parent=True,
-                type_="simple-entity",
-            ),
-            *(
-                _generate_date_range_entities(code_slot_num, parent=True)
-                if advanced_date_range
-                else ()
-            ),
-            *(_generate_dow_entities(code_slot_num, parent=True) if advanced_day_of_week else ()),
-        ]
-    )
-    return entities
-
-
-def _generate_parent_view_inner_card(
-    entities_card: MutableMapping[str, Any], parent_pin_entity_id: str | None
-) -> MutableMapping[str, Any]:
-    """Build the parent-view inner card, adding masked PIN markdown when needed."""
-    if parent_pin_entity_id:
-        return {
-            "type": "vertical-stack",
-            "cards": [
-                entities_card,
-                {
-                    "type": "markdown",
-                    "content": (
-                        f"{{% set pin = states('{parent_pin_entity_id}') %}}"
-                        "**PIN:** "
-                        "{% if pin not in ['unknown', 'unavailable', 'None', ''] %}"
-                        "Slot occupied"
-                        "{% else %}"
-                        "Empty"
-                        "{% endif %}"
-                    ),
-                },
-            ],
-        }
-    return entities_card
-
-
-def _generate_parent_view_card_ll_config(
-    code_slot_num: int,
-    advanced_date_range: bool,
-    advanced_day_of_week: bool,
-    parent_pin_entity_id: str | None = None,
-) -> MutableMapping[str, Any]:
-    """Build the parent-view conditional card for a child lock code slot.
-
-    Shows parent's settings alongside child's status when override_parent is off.
-    When parent_pin_entity_id is provided, the PIN row is replaced with a
-    read-only markdown card that indicates whether the slot is occupied.
-    """
-    entities = _generate_parent_view_entities(
-        code_slot_num,
-        advanced_date_range,
-        advanced_day_of_week,
-        parent_pin_entity_id=parent_pin_entity_id,
-    )
-
-    entities_card: MutableMapping[str, Any] = {
-        "type": "entities",
-        "show_header_toggle": False,
-        "state_color": True,
-        "entities": entities,
-    }
-
-    return {
-        "type": "conditional",
-        "conditions": [
-            _generate_state_condition(
-                code_slot_num, "override_parent", state="off", needs_type=True
-            )
-        ],
-        "card": _generate_parent_view_inner_card(entities_card, parent_pin_entity_id),
-    }
-
-
-def _generate_child_code_slot_dict(
-    code_slot_num: int,
-    advanced_date_range: bool,
-    advanced_day_of_week: bool,
-    parent_pin_entity_id: str | None = None,
-) -> MutableMapping[str, Any]:
-    """Build the dict for the code slot of a child keymaster lock."""
-    return {
-        "type": "grid",
-        "cards": [
-            _generate_header_ll_config(code_slot_num),
-            _generate_parent_view_card_ll_config(
-                code_slot_num,
-                advanced_date_range,
-                advanced_day_of_week,
-                parent_pin_entity_id=parent_pin_entity_id,
-            ),
-            {
-                "type": "conditional",
-                "conditions": [
-                    _generate_state_condition(code_slot_num, "override_parent", needs_type=True)
-                ],
-                "card": _generate_code_slot_conditional_entities_card_ll_config(
-                    code_slot_num, advanced_date_range, advanced_day_of_week, child=True
-                ),
-            },
-        ],
-    }
